requirements.txt
pandas
gspread
oauth2client
textblob
nltk
nrclex
joblib
google-cloud-translate
pyyaml